[GRPC_AgentClient] Local system id number
An AgentClient object must be initialized providing the id number of the
local system. This in order to embed a reference to the "sender" in the
RPC messages.

+++ plugins/agent_proxy/grpc/agent_proxy.h
@@ -138,7 +138,7 @@ private:
 
 	std::string GetNetAddress(int system_id) const;
 
-	std::shared_ptr<AgentClient> GetAgentClient(int system_id);
+	std::shared_ptr<AgentClient> GetAgentClient(uint16_t system_id);
 
 };
